Throws exceptions if ontology logically incoherent, fixes #40 Includes incoherent OPs in logical tests, fixes #104
Note these two tickets are coupled, so were tackled together.

To handle validation checks a new ReasonerHelper class was created.

This implements

 - inconsistency
 - incoherent TBox (ie unsat classes)
 - incoherent RBox (using probe classes)

Exceptions are thrown on encountering any of these.

Currently there is no way to allow the user to ignore exceptions;
this could in theory be added but would likely be a very bady idea