Update for Mac OS X 10.12
The following changes were required were required to run the robot
script on the Mac (OS X 10.12).

Note that I did not test this on any other systems but I added a test
to the script to see if it's running on a Mac and if so then the
following changes are implemented. If not running on a Mac then there
should be no change in behavior.

In the DIR assignment, putting double quotes around $0 didn’t work but
putting them around the ROBOT_SCRIPT variable did.