Fix query when chaining with input IRI
When chaining with an `--input-iri`, the `ontologyPath` variable in the state will be null. We currently try and guess the catalog path from this variable, but if the `ontologyPath` is null, creating a File object throws a null pointer exception. This PR adds a check to make sure the `ontologyPath` is not null before attempting to guess the catalog.