Allow to load extra commands from plugins.
When ROBOT's CLI is initialised, we search for implementations of the
`Command` interface offered as services, both in the system class path
and in any Jar file found in the following directories:

* the directory specified by the `robot.pluginsdir` system property;
* the directory specified by the `ROBOT_PLUGINS_DIRECTORY` environment
  variable;
* the calling user's `~/.robot/plugins` directory.

When an implementation is found, it is registered to the CommandManager,
allowing to use the command the same way as any built-in command.

When a command has been loaded from a Jar file from one of the
directories listed above (instead of from the system class path), its
name must be prefixed by the basename of the Jar file (without the .jar
extension) followed by a colon, for the command to be invoked from the
command line.

For example, if the file `my-plugins.jar` contains a command named
`hello`, that command must be invoked as `my-plugins:hello`.