Move PluginManagerIT to the robot-mock-plugin module.
It is counterintuitive that a test for the PluginManager, which is a
component of the robot-command module, shall be in a separate module.
But obviously we cannot test loading a plugin before the plugin itself
has been built, and we cannot build the plugin before robot-command
because the plugin itself is dependent on robot-command.

So we test the plugin manager during the integration testing phase of
the plugin; at this point all artifacts have been built (both the
robot.jar and the plugin jar).