Allow using a custom separator when merging annotations.
Add an optional argument to OntologyHelper.mergeExtraAnnotations to
allow the caller to specify the separator string to insert between
annotation values, when merging annotations together. The default
separator is a single space character.

This is a code-level feature only. The separator string is _not_
customizable from the command-line, when comment annotations are merged
as part of `--clean-obo merge-comments`.