Fix repair --merge-axiom-annotations.
The RepairOperation::mergeAxiomAnnotations() method only ever looks at
annotated axioms. As a result, it misses the case where the ontology
contains both an unnanotated axiom and one (or more) annotated
version(s) of that same axiom. For example:

  SubClassOf(UBERON:1 UBERON:2)
  SubClassOf(Annotation(rdfs:comment "A comment") UBERON:1 UBERON:2)
  SubClassOf(Annotation(rdfs:comment "Another one") UBERON:1 UBERON:2)

The last two axioms would be merged into a single one carrying the two
comments, but the first one would remain distinct:

  SubClassOf(UBERON:1 UBERON:2)
  SubClassOf(Annotation(rdfs:comment "A comment")
             Annotation(rdfs:comment "Another one")
             UBERON:1 UBERON:2)

We fix that by rewriting the method to forcefully look at _all_ axioms,
whether they are annotated or not.

Theoretically, this is sligthly inefficient as axioms that are not
annotated and that do not have any annotated duplicate would still be
removed and then added back. However I do not think that a more
elaborated approach would yield enough performance gain to be worth the
increased complexity.