Update test case for repair --merge-axiom-annotations.
Update the test case for `repair --merge-axiom-annotations` so that it
includes an instance of an unannotated axiom associated with annotated
duplicates.

This requires changing the format of the test case file from RDF/XML to
OFN, since RDF/XML does not allow to represent both an unannotated axiom
and an annotated version of the same axiom.