Allow interpolation of placeholders in annotation values.
This commit adds the possibility of using the `annotate` command to
inject ontology annotations with values derived from the ontology
itself, by substituting placeholders (of the form `%{NAME}`) within the
values provided on the command line.

For now, only two placeholders are supported (`%{ontology_iri}` and
`%{version_iri}`). For that reason, the implementation is very basic,
and is merely a couple of search-and-replace operations.