Include line separator when reading quries from Jar resources.
When we read a default query string from a resource embedded within the
Jar file, make sure to explicitly include a line separator between each
line (the line terminator is *not* included in the value returned by
BufferedReader#readLine()). Otherwise the entire query is stored as a
single line, which prevents its from being analyzed correctly by the
SPARQL parser.