Include SubObjectPropertyOf axioms when reducing.
When evaluating redundant SubClassOf axioms for the Reduce operation,
include SubObjectPropertyOf axioms in the graph, so that if we have

  SubClassOf(C, P1 some D)
  SubClassOf(C, P2 some D)
  SubObjectPropertyOf(P2, P1)

then the first SubClassOf axiom can be considered redundant and removed.

Also explicitly include SubPropertyChainOf axioms, so that similarly
redundant axioms when a property chain is involved can be removed as
well.