Add --include-subproperties option to 'reduce' command.
Do not unconditionally take into account SubObjectPropertyOf axioms when
evaluating redundancy in the context of a Reduce operation. Only do so
when this is expressly asked for with a new command-line option, to
avoid breaking existing pipelines.

(For the record, I believe this should be enabled by default, and that
"existing pipelines" could use an option to disable it instead.)