Partially fix prefix injection when writing to OBO.
When writing to OBO, prefixes added with `--add-prefix` are ignored.
This is because until recently, the OBODocumentFormat was not a
prefix-supporting format, and so there is nothing in the OWLAPI to
automatically take care of providing the prefixes the document may
contain into the OWL2OBO translator.

This commit fixes the issue at least for the case where the
`--clean-obo` option is used. In that case, we are instantiating the
OWL2OBO translator ourselves, so we can give it access to the prefixes
from the document.

For the case where `--clean-obo` is not used, the fix needs to happen in
the OWLAPI. In that case, we are relying on the ontology manager's save
methods, and we have no way to reach the OWL2OBO translator that is, in
fine, used by those methods when writing to OBO. (It would be possible
to always bypass the manager when writing to OBO, regardless of the
`--clean-obo` option, but that would be a more invasive fix).